requirements.txt
# Core
pyspark==3.5.0
delta-spark==3.0.0

# AI / LLM
anthropic==0.28.0
langchain==0.2.0
langchain-anthropic==0.1.15
langchain-community==0.2.0
openai==1.30.0

# Vector DB
chromadb==0.5.0

# API
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
httpx==0.27.0

# Pipeline
kafka-python==2.0.2

# Orchestration
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.14.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Linting
ruff==0.3.0
